feat(#1104): serve home composer at `/`, gate `/sequences/new` to signed-in users
Anonymous visitors land on the root with the same composer + showcase
instead of being redirected to `/sequences/new`. The logged-in alias stays
for sidebar / copy-sequence entry points.
